chore: flattens insert and upsert

+++ packages/db-postgres/src/upsertRow/index.ts
@@ -1,28 +1,10 @@
 /* eslint-disable no-param-reassign */
-import { Field } from 'payload/types';
-import { SQL, and, eq, inArray } from 'drizzle-orm';
-import { GenericColumn, PostgresAdapter } from '../types';
+import { and, eq, inArray } from 'drizzle-orm';
 import { transform } from '../transform/read';
 import { BlockRowToInsert } from '../transform/write/types';
 import { insertArrays } from '../insertArrays';
 import { transformForWrite } from '../transform/write';
-
-type Args = {
-  adapter: PostgresAdapter
-  data: Record<string, unknown>
-  fallbackLocale?: string | false
-  fields: Field[]
-  locale: string
-  path?: string
-  tableName: string
-  upsertTarget?: GenericColumn
-} & ({
-  where: SQL<unknown>
-  id: never
-} | {
-  id: string | number
-  where: never
-})
+import { Args } from './types';
 
 export const upsertRow = async ({
   adapter,
@@ -31,6 +13,7 @@ export const upsertRow = async ({
   fields,
   id,
   locale,
+  operation,
   path = '',
   tableName,
   upsertTarget,
@@ -46,23 +29,29 @@ export const upsertRow = async ({
     tableName,
   });
 
-  const target = upsertTarget || adapter.tables[tableName].id;
-
   // First, we insert the main row
   let insertedRow: Record<string, unknown>;
 
-  if (id) {
-    rowToInsert.row.id = id;
-    [insertedRow] = await adapter.db.insert(adapter.tables[tableName])
-      .values(rowToInsert.row)
-      .onConflictDoUpdate({ target, set: rowToInsert.row })
-      .returning();
+  if (operation === 'update') {
+    const target = upsertTarget || adapter.tables[tableName].id;
+
+    if (id) {
+      rowToInsert.row.id = id;
+      [insertedRow] = await adapter.db.insert(adapter.tables[tableName])
+        .values(rowToInsert.row)
+        .onConflictDoUpdate({ target, set: rowToInsert.row })
+        .returning();
+    } else {
+      [insertedRow] = await adapter.db.insert(adapter.tables[tableName])
+        .values(rowToInsert.row)
+        .onConflictDoUpdate({ target, set: rowToInsert.row, where })
+        .returning();
+    }
   } else {
     [insertedRow] = await adapter.db.insert(adapter.tables[tableName])
-      .values(rowToInsert.row)
-      .onConflictDoUpdate({ target, set: rowToInsert.row, where })
-      .returning();
+      .values(rowToInsert.row).returning();
   }
+
 
   let localeToInsert: Record<string, unknown>;
   const relationsToInsert: Record<string, unknown>[] = [];
@@ -104,14 +93,21 @@ export const upsertRow = async ({
   let insertedLocaleRow: Record<string, unknown>;
 
   if (localeToInsert) {
+    const localeTableName = adapter.tables[`${tableName}_locales`];
+
     promises.push(async () => {
-      [insertedLocaleRow] = await adapter.db.insert(adapter.tables[`${tableName}_locales`])
-        .values(localeToInsert)
-        .onConflictDoUpdate({
-          target: [adapter.tables[`${tableName}_locales`]._locale, adapter.tables[`${tableName}_locales`]._parentID],
-          set: localeToInsert,
-        })
-        .returning();
+      if (operation === 'update') {
+        [insertedLocaleRow] = await adapter.db.insert(localeTableName)
+          .values(localeToInsert)
+          .onConflictDoUpdate({
+            target: [localeTableName._locale, localeTableName._parentID],
+            set: localeToInsert,
+          })
+          .returning();
+      } else {
+        [insertedLocaleRow] = await adapter.db.insert(localeTableName)
+          .values(localeToInsert).returning();
+      }
     });
   }
 
@@ -123,40 +119,42 @@ export const upsertRow = async ({
 
   if (relationsToInsert.length > 0) {
     promises.push(async () => {
-      // Delete any relationship rows for parent ID and paths that have been updated
-      // prior to recreating them
-      const localizedPathsToDelete = new Set<string>();
-      const pathsToDelete = new Set<string>();
+      if (operation === 'update') {
+        // Delete any relationship rows for parent ID and paths that have been updated
+        // prior to recreating them
+        const localizedPathsToDelete = new Set<string>();
+        const pathsToDelete = new Set<string>();
 
-      relationsToInsert.forEach((relation) => {
-        if (typeof relation.path === 'string') {
-          if (typeof relation.locale === 'string') {
-            localizedPathsToDelete.add(relation.path);
-          } else {
-            pathsToDelete.add(relation.path);
+        relationsToInsert.forEach((relation) => {
+          if (typeof relation.path === 'string') {
+            if (typeof relation.locale === 'string') {
+              localizedPathsToDelete.add(relation.path);
+            } else {
+              pathsToDelete.add(relation.path);
+            }
           }
+        });
+
+        if (localizedPathsToDelete.size > 0) {
+          await adapter.db.delete(adapter.tables[`${tableName}_relationships`])
+            .where(
+              and(
+                eq(adapter.tables[`${tableName}_relationships`].parent, insertedRow.id),
+                inArray(adapter.tables[`${tableName}_relationships`].path, [localizedPathsToDelete]),
+                eq(adapter.tables[`${tableName}_relationships`].locale, locale),
+              ),
+            );
         }
-      });
 
-      if (localizedPathsToDelete.size > 0) {
-        await adapter.db.delete(adapter.tables[`${tableName}_relationships`])
-          .where(
-            and(
-              eq(adapter.tables[`${tableName}_relationships`].parent, insertedRow.id),
-              inArray(adapter.tables[`${tableName}_relationships`].path, [localizedPathsToDelete]),
-              eq(adapter.tables[`${tableName}_relationships`].locale, locale),
-            ),
-          );
-      }
-
-      if (pathsToDelete.size > 0) {
-        await adapter.db.delete(adapter.tables[`${tableName}_relationships`])
-          .where(
-            and(
-              eq(adapter.tables[`${tableName}_relationships`].parent, insertedRow.id),
-              inArray(adapter.tables[`${tableName}_relationships`].path, Array.from(pathsToDelete)),
-            ),
-          );
+        if (pathsToDelete.size > 0) {
+          await adapter.db.delete(adapter.tables[`${tableName}_relationships`])
+            .where(
+              and(
+                eq(adapter.tables[`${tableName}_relationships`].parent, insertedRow.id),
+                inArray(adapter.tables[`${tableName}_relationships`].path, Array.from(pathsToDelete)),
+              ),
+            );
+        }
       }
 
       insertedRelationshipRows = await adapter.db.insert(adapter.tables[`${tableName}_relationships`])
